Events: always scroll the current message list to the bottom on resize or channel switch
This is distinct from the scrolling behavior we get when messages
arrive.  That behavior is a result of a relative change in viewport
offset caused by wrapping the latest message in a visibility request.
This, on the other hand, is necessary because if the user resizes the
terminal the visible message might still be visible even if it is not at
the bottom!  Brick has done its job in keeping the message visible, but
we need to ensure that it is actually at the bottom, too.  So we have
this step on channel switch and resize.