Add ClientMessage and ClientPost type information, use to improve date transitions (fixes #53)
This change adds ClientMessageType for distinguishing between different
kinds of ClientMessages (which we now use, e.g., to improve date
transition rendering).  It also adds a new PostType field in lieu of the
booleans we had before for representing the different higher-level kinds
of posts: emotes, leaves/joins, etc.

This change also adds a new theme attribute for date transitions.