Break the interior of inline code sequences. Fixes #69
It correctly handles multiple spaces, as well, but non-space whitespace
(such as tabs) doesn't get handled specially, so that might produces some
anomalous behaviour. I currently believe that's enough of an edge case
that we might not bother considering it, but we may want to consider it
in the future.