Stop adding ourselves to joined channels (fixes #110)
Prior to this change, when we joined a channel we would go ahead and add
the channel to our known channel list. This was necessary because it was
our only opportunity to set up our channel metadata.

However, recently we started handling "I got invited to a channel"
messages from the server. These are sent when other users add us to
channels, and when we get those, it's a great time to go fetch the
channel metadata and set it up.  However, we *also* get those when we
join a channel ourselves! So prior to this change we'd add join a
channel, add the channel to our known channel list, then get a "you got
added to a channel" message from the server and do it all over again -
which was what resulted in adding the channel to our list twice.

The fix for this behavior was to make it so that when we join a channel,
we don't go ahead and add the channel to our channel list: we rely on
the server to send us a channel invite notification and we do it at that
time.  But that behavior reveals a server misbheavior that we weren't
prepared to handle: we'd get a "user was added to channel" message for
the new channel *before* we got our invite, and the 'addMessage'
function responsible for handling incoming messages would cause a
runtime exception trying to find the channel to add the message to
because we wouldn't have had a chance to register it!

So the "fix" for this is to make 'addMessage' just ignore a new message
when the channel ID for the new message isn't one we know about.
Admittedly this isn't great, but the only case when this should be
necessary is the one I've described above. We could always start logging
dropped messages in 'addMessage' if we suspect that this is being
overzealous.