Make the message preview scroll to track the input row position (fixes #120)
This change was made specifically to have to avoid adding manual
controls for scrolling the preview viewport.

Just prior to rendering an input message, we insert a fake "cursor
sentinel" character into the input at the end of the line currently
being edited.

This cursor character is not rendered by the Markdown-to-widget
conversion process; its presence is used to insert a brick "visibility
request" at the location of the cursor character when it is removed.
Since the character is still present in the markdown source during
pattern-recognition (such as for emojis), the character has to be
explicitly culled just prior to any matching that takes place. As a
result, any places where we looked at Text values in the Markdown code
now need to first call 'removeCursor' on those before doing things like
'isPrefixOf'. This does make the Markdown code a bit more brittle, but
it's not terribly hard to take this all out again if we decide we don't
want it and I think the usability improvement that we get for this is
worth the hack.