Fix a bug that caused webhook messages to not be considered unread
This change makes it so that webhook messages (which are "from" the user
who created the webhook) to be considered not "from" that user. This is
important because it means messages of this type will be considered "new
messages" in the channels where they appear even though they are from
your account. We need to check the username override because messages
from your account should be ignored when deciding whether a channel has
new messages (since you sent them; you know they're there).