Remove the old dark theme in favor of the "dark-clear" theme
This change makes it so that the dark theme works better by default on a
wider range of terminal color schemes. This is because we do not
explicitly set the background color, so if a terminal uses a light
background color by default this avoids e.g. getting black text on a
black background (as would have been the case with the old dark theme).
This theme also works better on terminals with dark themes when they
customize their background color, too.