Login: show authentication error on login failure
Since there are a number of ways that logins can fail (network errors,
credential errors, etc.), those are now collected into the
AuthenticationException sum type so that we can present these
appropriately in the Login interface. Previously we would just show the
UI again without telling the user why.

In an upcoming patch this will be improved further by permitting the
user to enter the hostname. That will not only make it optional in the
configuration but will make hostname-related configuration errors easier
to observe and repair.