Instead of whitelisting server commands, unify them with ordinary ones (fixes #162)
This change makes it so that server commands are implemented in the same
way as local ones: using the Cmd type. The benefits of this are that:

 * we get help output for free
 * places (such as completion) where we had to make special
   considerations for the server whitelist got generalized
 * we also get to do client-side validation of the input if desired

The execMMCommand routine got simpler, too, since it always runs what it
is told rather than checking the whitelist. Instead of an explicit
whitelist, we now have an implicit one: supported server commands have a
Cmd in the global commandList and unsupported ones do not.