Mark channels as "ChanLoadPending" when beginning an initial load
This change makes it so that channel-switching can avoid a duplicate
async scrollback fetch for unloaded channels. This is done by marking
the channel as "loading pending" so that we don't queue up more than a
single fetch for an unloaded channel. This is especially important to
avoid clobbering the "unread message cutoff"; an extra fetch for
scrollback clobbers that because it notices that there were no new
messages in the new fetch newer than what we already had. Naturally this
change also reduces network usage.