Use an STM TChan for async processing, add support for preemption in async scheduling
This change moves to using an STM TChan for the async work queue. This
is motivated by the desire to use STM.TChan.unGetTChan for placing
items at the front of channels. We use this functionality to support
preemption now: doAsyncWith now takes an AsyncPriority, either Preempt
or Normal, with Preempt placing the work item at the front of the queue
and Normal placing it at the back. All uses of doAsync(With) were
updated with a priority, with actions directly triggered by the user
usually using Preempt.

In particular, now that we have this feature, Town Square metadata and
message fetches on startup preempt all the other data loading that goes
on so we have much snappier startup behavior than before.