Log subprocess output and tell the user about it (fixes #175)
This change adds a subprocess logging thread and an associated TChan.
Send a ProgramOutput over that channel; if it has any non-empty stdout
or stderr, it will get logged to a tmpfile and a message will be posted
to the current channel telling the user what program just produced
output and where to find the log file containing that output.