State: initially set connection status to Connected
Previously we initially set the connection status to Disconnected and
let the initial websocket connection set it to Connected, but that no
longer works because we need to do some things during startup that
respect the connection status. Specifically, we updated the viewed state
of Town Square before we show the UI and that function refuses to run
when we're disconnected (to avoid exception noise). Because the
application doesn't define what it means to be started and disconnected
anyway, I think it's harmless to set this to Connected initially.
Everything else assumes that, anyway.