Fix new message transitions again
The new message cutoff timestamp indicates the time starting at which
messages have been unread in a given channel, so when inserting
transitions we need to consider messages *at* this time to be unread. I
broke the comparison for this and ended up showing the first unread
message as read as a result.