Split message-specific chunks of Types.hs apart
This cascaded into a second change that's good in the long-term, where
previously error/info messages had been created using a two-step
process of creating the message and then adding it to the chat state,
but we've had the bundled function to do it all at once
(`postErrorMessage`) for quite a while. For many modules, the only
message-related piece of data they needed was the `Error` or
`Informative` constructors to create error messages, so I've made
those uses use the `postErrorMessage` function instead, and added a
`postInfoMessage` function for the others. That means that, while many
files were touched in this commit, very few of them actually need
access to the internals of chat messages in any way.