Full overhaul: all stateful code lives in MH monad
This is a large, and highly experimental, change to the MatterHorn
codebase: this introduces a new monad, MH, which wraps both the
ChatState type as well as a continuation that describes what to do
when the event-handling finishes (i.e. whether to exit or not). This
is a change that has numerous advantages and disadvantages, but makes
certain kinds of tedious state-passing---a regular feature of MH
internals---much easier to manage. There are still a few standing API
questions, and several of the functions were not converted as nicely
as they could have been (e.g. cheating by accessing the full state
with `use id` instead of accessing only the needed variables.)