handleInputSubmission: clean up before handling the input
Previously we enough that it made sense to clean up after dealing with
the input because it is a natural ordering of operations: get input, act
on input, clean up.

But this broke things because then if the "act on input" step changed
the state, then the "clean up" step might clean up the wrong part of the
state -- and indeed it did, in the case of commands like `/focus`. That
command switches channels, and channel switches update the editor state.
This had the effect of leaving the `/focus` command in the editor for
the previous channel rather than clearing it as expected.