updateViewedIO: always Preempt other async work
This change makes it so that, effectively, at startup, Town Square does
not stay in an unread state, thus blocking the operation of M-a. Instead
we preempt async work when updating the channel state (which should
already have been the case with this particular API call) so that other
work doesn't interfere with the timely update of channel last-viewed
state.