Draw.Main: do not use a viewport for the channel list in ChannelSelect mode
This change avoids the use of "viewport" for the channel list when in
channel select mode. The rationale is that this mode does not need the
viewport because we don't give the user a way to navigate it anyway, but
also because the interface always keeps the list scrolled to the top
during ChannelSelect input. In addition the previous behavior meant that
entering ChannelSelect mode would cause the viewport offset to change
during input, and then when returning to normal mode this would be
evident in the channel list position, which was awkward.