Generalize runLoggedProgram (fixes #232)
This change makes it so that runLoggedProgram can indicate to the
subprocess logger whether a subprocess misbehaved enough to justify
telling the user. We do this by indicating whether the subprocess is
expected to produce stdout output (not all are). This change also makes
it so that the scripting infrastructure can now take advantage of
runLoggedProgram.