Add support for interactive URL-opening commands (fixes #266)
This is a mostly-complete patch to support interactive terminal browsers
in as the 'urlOpenCommand' setting. This works by pausing the
application event loop, restoring the terminal state, running the
interactive opener, and then bringing up both Vty and Brick once the
command has exited. This required refactoring openURL so that we can
run the opener either asynchronously (the old way) if the opener is not
interactive, or synchronously with application shutdown if it is.

This change also adds a config setting, urlOpenCommandIsInteractive
(bool), which the user must use to indicate the opener behavior. It
defaults to False.