Refactor tracking of New Messages indicators
This change adds a new type, NewMessageIndicator, with three
constructors:

 * Hide: do not show an indicator.
 * NewPostsStartingAt t: all posts in the channel with creation
   timestamps starting at 't' are new.
 * NewPostsAfterServerTime t: all posts in the channel with creation
   timestamps after 't' are new.

We have two modes with timestamps because when we initially get channel
metadata, the server metadata provides a last-viewed timestamp that is
*equal to* the creation timestamp of the most recently viewed post. In
that case, all posts created *after* that time are new. And when
rendering this case, we must put the indicator just *after* the post
with that creation timestamp.

In the case where new posts arrive after startup, we set the cutoff
*while handling new messages*, and instead of going and finding the most
recent message and using its timestamp, we use the timestamp of the
message in hand to set the cutoff. In this case, we want to consider
messages *starting* with this one to be new, so we use the "StartingAt"
constructor here. This is an optimization.

This also changes the notification logic so that new messages indicators
are no longer considered when deciding whether to notify the user with
NotifyUser.