Fix regression with user notifications on activity in the current channel
This change makes it so that we potentially take multiple actions
depending on the author of incoming messages and their relationship to
the current channel. In particular this change makes it so that in
addition to the previous behavior, we NotifyUser if any messages arrive
in the current channel that don't originate from the current user.