Draw.Main: generalize channel header drawing
This change fixes a bug where if a direct channel had a topic string
set, we didn't render the user info in the header: we rendered the
internal direct channel name ("id__id").

This change also generalizes the channel header rendering so that all
components of the header -- username, channel name, topic string -- are
treated as Markdown. Previously we only treated the topic this way when
it was being rendered in a non-direct channel header, and that led to
this behavior and other inconsistencies. This also means that now,
usernames and channel names mentioned in headers will use the same
colors when they are mentioned elsewhere. This change also removed the
user sigils from usernames in direct channel headers because 1) they're
redundant with the sigils in the sidebar and 2) they were breaking
Markdown username coloring.