Remove ___N combinators
This change removes the ___1, ___2, and ___3 combinators and replaces
them with a more idiomatic, direct style. This change may seem like a
bit of an abstraction regression, but the way we typically handle API
problems like this is to abstract over interstitial arguments using
a reader monad. In Matterhorn there are many places where calling
Mattermost API functions with a session and team make for verbose calls
and bulky lambdas, and the more idiomatic way to handle this problem is
to make a new wrapper API using a reader monad. This change backtracks
us to the older API so that we can upgrade it to a reader monad more
gracefully in the future. So later, instead of calling e.g.

  mmJoinChannel :: Session -> TeamId -> ChannelId -> IO ()

we'd call

  mmJoinChannel :: ChannelId -> MM ()

and

  runMM :: ChatState -> MM a -> IO a