Make channel_viewed, channel_updated events reload channel metadata
This change makes it so that for either of these websocket events, our
reaction is the same: reload the channel metadata.

This change also makes it so that when a topic change occurs, we no
longer set the channel topic when the cosmetic message arrives. Instead
we rely on channel_udpated to fire so we know to go fetch the metadata.
Before, we were even doing double duty: we set the topic when the change
command was being sent and we also set it when the notification message
arrived.