refreshChannels: handle new channels
This change makes it so that refreshChannels is capable of dealing with
channels that are new to this client since the last reconnection. This
change also restricts initializeState to start with only Town Square at
startup, both to satisfy the channel zipper's non-empty requirement and
to facilitate further testing as this branch develops.