Make handleNewChannel canonical for choosing channel names
Previously callers duplicated logic for choosing channel names. Now,
handleNewChannel always does the right thing and callers have less work
to do. In addition, because the new user-fetching is deliberately
restricted to only users in the current team, this meant that
handleNewChannel has a new responsibility: it might get a (DM) channel
associated with a user that is no longer in the current team. Since the
user load doesn't get those users, we need to get it separately and
defer the channel registration until we have the data.