asyncFetchAttachments: don't add duplicate attachments
This change fixes a bug where if a message had attachments and then was
edited, we'd re-fetch the attachment info and add it blindly, leading to
duplicates showing in the UI. Instead, when adding attachments we skip
them if they're already in the message. This change is required because
asyncFetchAttachments used to only be used to fetch attachments the
first time we got a message, but it is now also used to update
attachments on message edit.

One omission from this function is that it doesn't consider removed
attachments, but I'm not sure that matters anyway: as best I can tell,
the web UI doesn't let you remove attachments anyway.