Use message characteristics to determine layout
Currently the message layout is the same for all messages. For block
messages or messages that span multiple lines this means the entire
message is indented after the end of the user's name. This commit
changes the message layout so that block type messages or messages
spanning multiple lines are laid out vertically so the message appears
under the user's name rather using a horizontal lay out. The
horizontal layout is still used for single line messages.