setLastViewedFor: ignore requests for nonexistent channels
Previously this function treated a request to update a channel timestamp
for a channel *absent* from the state as a channel *invite*. Under some
conditions is isn't totally crazy to do, since it means we attempt to
become consistent by assuming requests to update the viewed timestamp
for unknown channels means we somehow missed adding it to the channel
list.

However, it *is* the case that after deleting channels we will still get
WMChannelViewed websocket events for the deleted channel, and we were
previously turning those into a call to handleChannelInvite which was
the absolute opposite thing we wanted!

This change stops doing that, under the assumption that if turning these
spurious events into invite handling was necessary, there is a better
way to handle that situation that doesn't get us into trouble
misinterpreting timestamp update requests.