runLoggedCommand: run asynchronously (fixes #270)
This change makes it so that instead of running subprocesses
synchronously and returning their output, runLoggedCommand accepts an
optional MVar into which the output will be placed when it is ready. If
no MVar is given, the output will not be made available to the caller.
In either case, the subprocess will be run asynchronously and the call
to runLoggedCommand will return immediately.