Make editing indicators configurable
This change adds a configuration setting, 'showOlderEdits' (default
True), which determines whether "(edited)" is shown for edited posts
that are not edited recently (i.e. before the channel was last viewed).
This way they can be distinguished from edits that *are* recent and can
be turned off if they're deemed noisy.