Make Config aware of its own location when possible
This change removes the unpleasant API I had introduced for tracking
the path that we used to load the configuration. We need to track that
path because it's used to resolve relative file references in the
configuration after configuration load time. This change is intended
to address Getty's review comments and improve on the previous API by
making the Config aware of its own location (optionally).