handleNewChannel: ignore channels that already exist in the state (fixes #327)
This change makes it so that handleNewChannel does nothing if it gets
told to handle a new channel that already appears in the state. This can
happen when we get told multiple times about a channel, e.g., the local
user creates a new ad-hoc group channel and the server also tells us it
has been created. Both events need to update the state but if one of
them clobbers the other, the channel gets stuck in a bad (unloaded)
state.