addMessageToState: automatically show hidden group channels (fixes #326)
If we receive a message for a channel we don't know about, we now fetch
the channel metadata for the channel and add it to the state.
Specifically, if that channel is a Group channel, that means it is a
group channel that has been set to Hidden by the user. So we now issue a
preference change to show it again and add it to the state. This brings
our behavior in line with that of the web client.