spaces in user statuses; more username color-hash control
Previously, `colorUsername` in `Themes.hs` had one argument---a
username---and it would try to filter out unnecessary bits (like
leading sigils or status indicators) before hashing that username for
color. This was brittle, and broke when adding leading spaces before
user statuses. So instead, `colorUsername` now takes two arguments:
the raw username to hash on, and the displayed text we'd like.

This will be especially useful when we switch to using richer display
names as described in #265: we can still hash on the username but
nevertheless display the richer name, so name coloration will be
stable between usernames and display names.

This change also puts a space before user statuses, and does a minor
refactor of the type of channel sigils that turns out to simplify its
use a lot.