KeyEvents: make key event name mistakes impossible
This change removes the duplication of string literals to refer to key
events, thus avoiding mistakes in duplication that could cause
round-tripping of key events and names to break. It depends on the
assumption that the "allEvents" list lists all possible events that we
would want to name, but right now that list doesn't actually contain all
possible events so it's possible I misunderstood the purpose of that
list.

Also, this change maintains the case expression that maps a name to each
key event to ensure that as new events are added, we at least get
compiler warnings about having failed to assign names to them.