Draw.Main: reinstate passing HighlightSet to places that need it
This change reinstates the previous pattern of passing the HighlightSet
to the spots that need it to ensure that the set is shared rather than
recomputed. On servers with sufficiently many users, the cost to build
the set numerous times introduced significant input latency due to
costly redraws that weren't sharing the sets.