Use UserRef instead of UserId for LinkChoice values
Using a UserId means that the only links which can appear in the URL
select view are links from specific concrete users, but this is a
problem for bots (which post _as_ users but with an overloaded
username). We had started to be too specific and ended up filtering
out all bot-posted messages from URL select consideration, so this
fixes that by allowing LinkChoice values to contain a generic UserRef
value instead.