Start work on an overlay for displaying lists of users
To begin with, this will replace the dump-usernames-to-channel
behavior of /members with an overlay that displays a rich scrollable
list of users, complete with extra metadata. In the future, this same
list can be used for any situation where we want to browse lists of
users, including finding users on a team or what-have-you.

Importing missing features as of this commit include:
- actually implementing scrolling, which is not yet present
- implementing status lookup