Types: stop exposing csUsers lens
This change moves some functions from State to Types and adds
MH-specific wrappers around some Types.Users functions on the Users data
structure. It also stops exporting the csUsers lens to force callers to
go through the interface in Types to manipulate the user state. This
change also renames updateStatus to updateUserStatus and simplifies the
implementation of the status update thread.