State: generalize handleNewUser to fetch information for a collection of users
This change generalizes handleNewUser (renaming it to handleNewUsers in
the process) so that it can fetch the information for one or more user
IDs. This is then used to optimize the fetching of post authors in
addObtainedMessages: there, when posts are added, we check to see
whether we know about the authors. If we don't, all of the unknown users
are submitted to handleNewUsers in a single batch.