addObtainedMessages: avoid all reply parent fetches by using post map
It turns out that when you ask for a page of posts, the server
automatically gives you those posts *and* the posts in threads
associated with those posts. We were ignoring those extra posts in the
response because we were only looking at the post ordering ID list, and
ignoring extra posts in the post map. This meant that we were queueing
up a fetch for a post each time we encountered a reply to it! So we were
fetching the same posts multiple times when we could have avoided that
entirely.

This change goes ahead and updates the csPostMap, which is responsible
for storing such posts, with the data that the server gave us before we
add the individual posts to the state. This way, addMessageToState will
never fetch reply parents in those cases, drastically improving startup
performance on slow connections or in channels with lots of threading
activity.