UserListOverlayState: add UserSearchScope type and field
This change makes it so that instead of storing the kind of user list
in the Mode value, we store it in the UserListOverlayState itself.
The thing we store there is called a UserSearchScope. It has two
constructors: ChannelMembers <channel_id> and AllUsers. The former means
"search for and show users restricted to searching within the specified
channel" and the second will probably mean "show users found within the
current user's team" once it's done (although I'm not sure if the "team"
bit should be there).