UserListOverlay: prefetch more results as the user's cursor nears the end of the list
This change makes it so that when no search string is in effect (the
only case where we can paginate the results) and the user moves the
cursor to within a small distance from the end of the list of users, we
go ahead and fetch the next page of results and append them to the list.
See the docs for prefetchNextPage for details on how we decide when to
do this.