UserListOverlay: add handlers for Enter keypresses
This also adds a new key event, ActivateListItemEvent, for when a user
wants to "activate" a selected list item (e.g. by pressing Enter on it
to take some action).

With this change the user list shown by /members triggers a DM chat for
the selected user on Enter keypress.