UserListOverlay: make enter handlers return whether they fired
This change fixes a bug where filtering the current user from the list
of results caused trouble in page prefetching. The issue is that
filtering the current user from the list of results breaks pagination
logic that we use to figure out whether we should ask for more results.
Rather than make that logic trickier, I just punted and make it so that
the current user *does* appear in the list, *but* the Enter keypress
handler for each use of the overlay gets to decide what to do with that
entry when it is activated. This is more general and doesn't bake that
assumption into the API, although it does mean that the handlers for
Enter keypresses have to do a little more work. To ensure that those
events don't close the overlay, the Enter handlers now return a bool
indicating whether they did anything useful, with False meaning that the
overlay should stay open and True meaning that it can be closed.