Bugfix: replies now indicate the correct root post ID
Previously, a message sequence

  A
  B
  C

with B replying to A and C replying to B would create two *separate*
threads:

  A
  B

and

  B
  C

which is not how the web client works, and breaks up conversations
otherwise intended to be part of the same thread. Now, we do the right
thing: when posting a reply message, its "root ID" is not the ID of its
immediate parent, but the ID of the thread root post, as intended.