Catch IO errors during login and display nicer errors
This fixes #358, but might require more work in the future: right now,
this code assumes that the `IOError` in question will only ever be
raised in response to the network being down. That might not be
correct, and we should do further testing to find out other situations
under which the `IOError` in question might be raised, and refine our
error messages appropriately!