UserListOverlay: disable pagination, always use /users/search
This change temporarily disables the use of /users to get user search
results. This is because that endpoint does not allow us to hide deleted
users from its results, while /users/search does -- which means that the
results will vary undesirably between the two. Our user list overlay
used to use the former when the user had not entered a search string and
switched to the latter when they had entered something. This meant that
when not searching, the user could freely request more pages of the user
list. But due to the deleted users appearing there, it gave the wrong
results.

This change uses an unfortunate but necessary hack: when the user has
not entered a search string, we submit a user search request with a
string containing exactly one space. According to the developers and
empirical testing, this is the equivalent to making a request to /users
as if the user had not entered a search string. The empty string is not
valid, and the upstream developers have acknowledged that it should be
fixed. In the mean time, we use this hack.

The other important change is that pagination requests to get more
users are now disabled. The only endpoint that supported pagination at
all was /users, so the logic for page prefetching and some of the
corresponding book-keeping and UI hints were disabled. However, note:
they were not *removed*, since we hope one day to reinstate that logic
once the server API improves.