Omit join/leave messages if the user wishes (fixes #363)
This is a purely display-oriented preference: we still internally
store the join/leave messages, and we also update this preference live
if it is updated from the web UI. There is not yet a way to set this
from Matterhorn, but we should add that soon.

This also introduces a new field to csResources, but I'd like to
eventually remove an existing field to make up for it. The new field
is a wrapped, rich UserPreferences type that right now reflects a
single preference we fetched from the server. We _also_ store a
sequence of Preference values, which we should move away from---all
that information (represented now as stringly-typed key/value pairs)
should eventually be moved into UserPreferences.

Note: I do not believe that we currently do fetch preferences on
reconnect, so if you disconnect a Matterhorn session, toggle the
preference, and reconnect, then we will not honor the preference until
you reconnect. This is also true for other preference-based settings
(like flagged messages) so we should change this all in one swoop.