updateSelectMatch: fix C-n/C-p bug resulting from inconsistent data
This change makes it so that updateSelectMatch just uses whatever was
already in the user select list to determine the match index change.
This is how it should have been done already, and this change makes the
function simpler. The recent change to support nicknames revealed an
inconsistency bug where the data used to build the user match list
(elsewhere) could disagree with the extra user list logic that this
patch removed, leading to weird/broken cursor movements in the C-g list.