MMNames: distinguish between channel and user name mappings
This change addresses #383 by splitting up the name-to-channel-ID
mapping into two mappings: one for channel names and one for user names.
This is necessary because these names form two namespaces and we were
collapsing them into one, causing conflicts when a channel and user
shared the same name.

This change added new functions, channelIdByUsername and
channelIdByChannelName, to let the caller specify which namespace to
search. It also adjusted channelIdByName to return a *list* of matches
by searchin both namespaces, to let the caller decide how to handle
clashes. That function also uses a sigil to decide which namespace to
ignore to let the user input disambiguate.