Overhaul channel selection mode data structure management
Overall this patch is aimed at ensuring that users and channels are
treated separately for matches in channel selection mode. But it touched
a lot of stuff having to do with this to make it work: now, instead of
just storing the textual match value and using that to look for the
"current" selected match, we store what *kind* of match it was. Then,
when rendering the match results, we compare the match (a MatchValue) to
the different sections of the sidebar to ensure that we don't confuse a
username match with a channel match.

This patch also got rid of some extra conversions between Hashmaps and
lists since the hashmap lookups were only ever needed at draw time, but
in results navigation we needed the original sorted results. This patch
avoids extra sorting operations whenever only the selection index
changes.

Finally, this patch makes channel selection mode use a viewport when
rendering the sidebar, and it also asks each channel group fetching
function to filter the fetched results so that we can optimally only
render enough sidebar to fill the screen. This manifests as a new
argument to e.g. getDmChannels of type (ChannelListEntry -> Bool). It's
necessary to do this kind of filtering in e.g. getDmChannels so that the
position of the "current" item in such lists can be used to get the
surroundings according to the screen height argument.

Overall, the sidebar rendering logic is nasty. This patch makes it
slightly nastier.