Config: add syntax directory setting
This change adds a new config field, configSyntaxDirs, and an associated
setting, syntaxDirectories, to allow the user to specify the directories
from which to load XML syntax definitions.

The setting is optional. If omitted, it defaults to the
"syntax" subdirectory of the XDG configuration directory (i.e.
~/.config/matterhorn/syntax) followed by the "syntax" directory that
is either adjacent to the running matterhorn binary (as in the bindist
case) or adjacent to the dist-newstyle directory (as in the development
case where the matterhorn binary is located deep within dist-newstyle in
the repository).