Add syntax map to ChatResources
This change adds a SyntaxMap field to ChatResources and then updates the
Markdown renderer to use it instead of Skylighting's default syntax map.
This change prepares us to use our own syntax map loaded from disk
rather than the one that we are about to stop using once we move to
skylighting-core.