Improve emote editing: remove and reinstate server-side formatting
This change makes it so that, prior to editing an emote, we strip the
server-side formatting. We do this because that formatting is *metadata*
that we use to determine that a post is an emote (along with other
metadata). However, it is also *formatting* that is undesirable for
Matterhorn users because editing it can break the message's emote
status. We remove the formatting when editing begins, and reinstate it
just prior to invoking the API endpoint to submit changes to the post.
This way, the server's view of the message is the same, and when we get
notified of the edit later, we still conclude that it's an emote and
render it correctly. In the mean time, the user doesn't have to
awkwardly edit the formatted version of the message.

This change also makes it so that when we preview such edits, they
*look* like emotes because we tell the preview formatter to use the type
of the edited message to influence how the preview is drawn.

All in all, it's a mess.