Message: add MessageId type to wrap PostId
Before this change, message selection mode was only capable of selecting
messages with PostIds. This was by design, but it made it impossible to
select messages of other types, which we want so that we can support
yanking (or performing other actions on) arbitrary messages.

This change introduces a MessageId type, so that messages created
internally can be assigned IDs (in this patch, UUIDs). This patch does
not actually start to manufacture such UUIDs, though; it just does the
plumbing to get the new MessageId constructors used in the right places
to preserve the previous functionality.