Add "message view" mode for viewing messages with long lines (fixes #377)
This change adds a new binding to message selection mode. The binding,
named "view-message", defaults to 'v'. When used, the binding opens the
selected message in a pop-up window by itself in which arrow keys and
paging keys can be used to view the message. This feature is for viewing
messages whose long code block lines have been truncated in normal
rendering. (We don't wrap code block lines because it often renders the
code block content invalid or confusing when read or used.)