Types: add a logging context to the MH monad
This change adds a "logging context" to the MH monad. The logging
context is information that should accompany a log message so that the
relevance of the message is clearer.

The context is optional and defaults to Nothing. The context can be set
with withLogContext and can be obtained with getLogContext.

This change does not actually add logging functionality.